Fix symmetric tensor multi-index reordering in canonicalization
This fixes a bug where not all multi-indices that are equivalent under symmetry were being sorted to the same canonical form. It also updates the creation of the mapping of storage indices to multi-indices to elide canonicalizing multi-indices of non-symmetric tensors. The latter is being done in the same commit as the former because the former increases compile-time complexity and the latter reduces the number of constexpr evaluation steps for what would otherwise be the worst cases when creating the mapping between storage and multi-indices: high rank non-symmetric tensors.